+++ NEWS.md
@@ -13,7 +13,7 @@ There was a significant improvement in the speed of calculation. Details in the 
 
 -   `plot.gips()` can get `type = "n0"`, which will plot the change of `n0` along the "MH" optimization. Handy for deciding of burn-in time;
 -   `find_MAP(optimizer = "MH")` tracks the `n0` along the optimization;
--   `summary.gips()` calculates Likelihood-Ratio test (in development, TODO: documentation, mention in vignette, LR for estimated mean);
+-   `summary.gips()` calculates Likelihood-Ratio test;
 
 
 # gips 1.2.2